+++ Sources/InstanaAgent/Utils/Extensions/Double+Misc.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+extension Double {
+    static func == (lhs: Double, rhs: Double) -> Bool {
+        return abs(lhs - rhs) < 0.000001
+    }
+
+    static func != (lhs: Double, rhs: Double) -> Bool {
+        return !(lhs == rhs)
+    }
+}